Support uint64 short URL sequence and extend Base58 ID format (#27)
Migrate counter interfaces and storage from uint32 to uint64, update Feistel ID generation to preserve 6-char compact IDs while appending a Base58 suffix as needed, extend valid short ID length to 12, and document the format in README.

+++ internal/testhelpers/id_generator.go
@@ -19,7 +19,7 @@ func NewIDGenerator(id string) *IDGenerator {
 }
 
 // Generate returns the configured fixed ID.
-func (g *IDGenerator) Generate(_ uint32) string {
+func (g *IDGenerator) Generate(_ uint64) string {
 	g.calls.Add(1)
 	return g.id
 }
